Extract path to separate module
