+++ lightweight/__init__.py
@@ -1,6 +1,7 @@
-from .site import Site, SitePath
+from .site import Site
 from .content import Content, markdown, render, sass
 from .files import paths
+from .path import SitePath
 from .template import template
 
 __version__ = '0.1.0.dev13'
